common/types/utils.tsを整理  (#95)
* chart関連の型移動

* button関連の型移動

* utilsファイルの残りの型移動

* sessionUserの命名変更
