+++ common/types/ui-elements/button.ts
@@ -1,0 +1,8 @@
+export type FormSubmitButtonProps = {
+  onClick: React.MouseEventHandler<HTMLButtonElement>;
+  label: string;
+};
+
+export type FormCancelButtonProps = {
+  onClick: React.MouseEventHandler<HTMLButtonElement>;
+};
