+++ common/types/ui-parts/layout.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export type ChildrenProps = {
+  children: ReactNode;
+};
