+++ features/skills/types/api.ts
@@ -1,9 +1,9 @@
 import { ApiOptions } from "@/common/types/api";
 import { StackRankings } from "../components/RankTable";
-import { sessionUser } from "@/features/sessions/types/session";
+import { SessionUser } from "@/features/sessions/types/session";
 
 export type FetchStackRankingsProps = {
   options: ApiOptions;
-  sessionUser: sessionUser;
+  sessionUser: SessionUser;
   setStackRankings: React.Dispatch<React.SetStateAction<StackRankings[]>>;
 }
