+++ features/users/contexts/UserFormContext.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useState, useMemo } from 'react';
-import { ChildrenProps } from '@/common/types/utils';
 import { InitialUserFormData } from '../functions/form';
 import { UserFormContextProps, UserFormDataParams } from '../types/context';
+import { ChildrenProps } from '@/common/types/ui-parts/layout';
 
 const InitialState: UserFormContextProps = {
   userFormData: InitialUserFormData,
